Remove tf BatchPolopt
