+++ src/garage/tf/__init__.py
@@ -1,1 +1,4 @@
 """Tensorflow Branch."""
+from garage.tf._functions import paths_to_tensors
+
+__all__ = ['paths_to_tensors']
