+++ src/garage/tf/algos/__init__.py
@@ -1,5 +1,4 @@
 """Tensorflow implementation of reinforcement learning algorithms."""
-from garage.tf.algos.batch_polopt import BatchPolopt
 from garage.tf.algos.ddpg import DDPG
 from garage.tf.algos.dqn import DQN
 from garage.tf.algos.erwr import ERWR
@@ -15,7 +14,6 @@ from garage.tf.algos.trpo import TRPO
 from garage.tf.algos.vpg import VPG
 
 __all__ = [
-    'BatchPolopt',
     'DDPG',
     'DQN',
     'ERWR',
